3.1: * Items inspect * Talents inspect * Updated PET_SPELLS packet * Enchantments don't screw up display of items * Frozen Power proc
git-svn-id: svn://svn.aspiredev.org/AspireCore/branches/3_1_0@1384 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/hearthstone-world/SpellHandler.cpp
@@ -382,7 +382,7 @@ void WorldSession::HandleAddDynamicTargetOpcode(WorldPacket & recvPacket)
 	uint64 guid;
 	uint8 counter;
 	uint32 spellid;
-	uint32 flags;
+	uint8 flags;
 	UnitPointer caster;
 	SpellCastTargets targets;
 	SpellEntry *sp;
